fix: refactor accounts code into new files.
